pyproject.toml
[project]
name = "microblog"
version = "0.1.0"
description = ""
authors = [
    {name = "Ruairi O'Donohoe",email = "58832267+ruairiodonohoe@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "djhtml (>=3.0.8,<4.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "email-validator (>=2.2.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "microblog", from = "src"}]


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
